move components to their own folder
